implemented all tests for db seeding
